- added dao and repository for Wish and Wishlist - fixed background in css

+++ src/main/java/com/example/digitalwishlist/model/user/UserDao.java
@@ -4,23 +4,29 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.util.Streamable;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
 
 @Service
 public class UserDao {
-  @Autowired
-  private UserRepository repository;
 
-  public void save(User user) {
+  private static UserRepository repository = null;
+
+  @Autowired
+  public UserDao(UserRepository repository) {
+    UserDao.repository = repository;
+  }
+
+  public static void save(User user) {
     repository.save(user);
   }
 
-  public void delete(User user) {
-    repository.delete(user);
-  }
-
   public void delete(String email) {
+    boolean exists = repository.existsById(email);
+    if (!exists) {
+      throw new IllegalStateException("User with email:  " + email + " does not exist");
+    }
     repository.deleteById(email);
   }
 
@@ -29,5 +35,39 @@ public class UserDao {
     Streamable.of(repository.findAll())
         .forEach(users::add);
     return users;
+  }
+
+  public List<User> getUser() {
+    return (List<User>) repository.findAll();
+  }
+
+  @Transactional
+  public void updatePassword(String email, String password) {
+    User user = repository.findById(email).orElseThrow(() -> new IllegalStateException(
+        "User with email:  " + email + " does not exist"));
+
+    if (password != null && !password.equals(user.getPassword())) {
+      user.setPassword(password);
+    }
+  }
+
+  @Transactional
+  public void updateFirstName(String email, String givenName) {
+    User user = repository.findById(email).orElseThrow(() -> new IllegalStateException(
+        "User with email:  " + email + " does not exist"));
+
+    if (givenName != null && !givenName.equals(user.getGivenName())) {
+      user.setGivenName(givenName);
+    }
+  }
+
+  @Transactional
+  public void updateLastName(String email, String lastName) {
+    User user = repository.findById(email).orElseThrow(() -> new IllegalStateException(
+        "User with email:  " + email + " does not exist"));
+
+    if (lastName != null && !lastName.equals(user.getLastName())) {
+      user.setLastName(lastName);
+    }
   }
 }
